refactor: Move isTooltipComptaible to Piece tooltip component
