+++ src/app/rundown-view/components/offsetable-piece/offsetable-piece.component.html
@@ -1,6 +1,6 @@
 <div class="label" [style.margin-left.px]="labelOffsetInPixels">
   <span #labelTextElement>{{ piece.name }}</span>
 
-  <sofie-piece-tooltip *ngIf="isTooltipCompatible" [piece]="piece" [durationInMs]="partDuration" [playedDurationForPartInMs]="playedDurationForPartInMs" [isMediaUnavailable]="isMediaUnavailable" [media]="media">
+  <sofie-piece-tooltip [piece]="piece" [durationInMs]="partDuration" [playedDurationForPartInMs]="playedDurationForPartInMs" [isMediaUnavailable]="isMediaUnavailable" [media]="media">
   </sofie-piece-tooltip>
 </div>
